Class with all the necessary commands for robot
Has angle turning, moving to a point (need definitions of points from vision team for calibration). Also needs kicker, which needs to be discussed with engineering team. Possible reverse movement, but need more information on conditions (should it decide by itself, or should there be a separate method for it). Will be implemented hardcoded check for boundaries.